requirements.txt
streamlit
pandas
google-auth
google-auth-oauthlib
google-cloud-bigquery
requests